change IDepositContract interface

+++ packages/loopring_v3/contracts/impl/BasicDepositContract.sol
@@ -71,18 +71,18 @@ contract BasicDepositContract is IDepositContract, ReentrancyGuard
     function deposit(
         address from,
         address token,
-        uint    amount,
+        uint96  amount,
         bytes   calldata /*auxiliaryData*/
         )
         external
         override
         payable
         onlyExchange
-        returns (uint actualAmount, uint tokenIndex)
+        returns (uint96 actualAmount, uint tokenIndex)
     {
         // Check msg.value
         if (isETHInternal(token)) {
-            require(msg.value == amount, "INVALID_ETH_DEPOSIT");
+            require(msg.value == uint(amount), "INVALID_ETH_DEPOSIT");
         } else {
             require(msg.value == 0, "INVALID_TOKEN_DEPOSIT");
              // Transfer the tokens from the owner into this contract
@@ -90,14 +90,14 @@ contract BasicDepositContract is IDepositContract, ReentrancyGuard
                 token.safeTransferFromAndVerify(
                     from,
                     address(this),
-                    amount
+                    uint(amount)
                 );
             }
         }
 
         uint exchangeBalanceBefore = exchangeBalance[token];
         // Keep track how many tokens are deposited in the exchange
-        uint exchangeBalanceAfter = exchangeBalanceBefore.add(amount);
+        uint exchangeBalanceAfter = exchangeBalanceBefore.add(uint(amount));
         // Make sure the total max amount per token in the exchange is capped
         require(exchangeBalanceAfter <= MAX_TOTAL_TOKEN_BALANCE, "MAX_AMOUNT_REACHED");
         exchangeBalance[token] = exchangeBalanceAfter;
@@ -118,7 +118,7 @@ contract BasicDepositContract is IDepositContract, ReentrancyGuard
         onlyExchange
     {
         // Keep track how many tokens are deposited in the exchange
-        exchangeBalance[token] = exchangeBalance[token].sub(amount);
+        exchangeBalance[token] = exchangeBalance[token].sub(uint(amount));
 
         // Transfer the tokens from the contract to the recipient
         transferOut(to, token, amount);
@@ -188,7 +188,7 @@ contract BasicDepositContract is IDepositContract, ReentrancyGuard
     function transferOut(
         address to,
         address token,
-        uint amount
+        uint    amount
         )
         internal
     {
